# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Snap is an open telemetry framework designed to simplify the collection,\nprocessing and publishing of system data through a single API.\n\nThis should demonstrate all the possible comment annotations\nthat are available to turn go code into a fully compliant swagger 2.0 spec",
    "title": "Snap Restful API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "name": "Snap Maintainer",
      "url": "http://snap.io",
      "email": "snap@intel.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/"
    },
    "version": "1.0.0"
  },
  "host": "127.0.0.1:8181",
  "basePath": "/v2",
  "paths": {
...
    "/tasks": {
      "get": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "lists a list of created tasks. An empty list returns if no created tasks.",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "$ref": "#/responses/TasksResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "creates a task based on the input. Othereise, an error returns if the input misses the required fields or is in a malformed format.",
        "operationId": "addTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Task",
            "description": "Create or update a task.",
            "name": "task",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/TasksResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "lists a task by the giving task id. Otherwise a not found error returns.",
        "operationId": "getTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "updates a task's state for the giving task id and the input state. An error occurs for any bad request.",
        "operationId": "updateTaskState",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "State",
            "description": "Update the state of a task",
            "name": "action",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/TaskResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "deletes a task for the giving task id. Note that only a stopped task may be removed. Otherwise, an error occurs.",
        "operationId": "removeTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/TaskResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/TaskErrorResponse"
          }
        }
      }
    },
    "/tasks/{id}/watch": {
      "get": {
        "consumes": [
          "application/json",
          "application/x-protobuf",
          "text/event-stream"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf",
          "text/event-stream"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "watches a task data stream for the giving task id. Otherwise, an error returns.",
        "operationId": "watchTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
    "CollectWorkflowMapNode": {
      "type": "object",
      "title": "CollectWorkflowMapNode defines the workflow type.",
      "required": [
        "metrics",
        "process",
        "publish"
      ],
      "properties": {
        "config": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "x-go-name": "Config"
        },
        "metrics": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/MetricInfo"
          },
          "x-go-name": "Metrics"
        },
        "process": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessWorkflowMapNode"
          },
          "x-go-name": "Process"
        },
        "publish": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PublishWorkflowMapNode"
          },
          "x-go-name": "Publish"
        },
        "tags": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "x-go-name": "Tags"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/scheduler/wmap"
    },
...
    "MetricInfo": {
      "type": "object",
      "title": "MetricInfo defines metric version.",
      "properties": {
        "version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Version_"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/scheduler/wmap"
    },
...
    "ProcessWorkflowMapNode": {
      "type": "object",
      "required": [
        "name",
        "process",
        "publish"
      ],
      "properties": {
        "config": {
          "description": "TODO processor config",
          "type": "object",
          "additionalProperties": {
            "type": "object"
          },
          "x-go-name": "Config"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "plugin_version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Version"
        },
        "process": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessWorkflowMapNode"
          },
          "x-go-name": "Process"
        },
        "publish": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PublishWorkflowMapNode"
          },
          "x-go-name": "Publish"
        },
        "target": {
          "type": "string",
          "x-go-name": "Target"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/scheduler/wmap"
    },
    "PublishWorkflowMapNode": {
      "type": "object",
      "required": [
        "name",
        "config"
      ],
      "properties": {
        "config": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          },
          "x-go-name": "Config"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "plugin_version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Version"
        },
        "target": {
          "type": "string",
          "x-go-name": "Target"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/scheduler/wmap"
    },
...
    "Schedule": {
      "type": "object",
      "title": "Schedule defines a scheduler.",
      "required": [
        "type",
        "interval"
      ],
      "properties": {
        "interval": {
          "type": "string",
          "x-go-name": "Interval"
        },
        "start_timestamp": {
          "x-go-name": "StartTimestamp"
        },
        "stop_timestamp": {
          "x-go-name": "StopTimestamp"
        },
        "type": {
          "type": "string",
          "enum": [
            "simple",
            " windowed"
          ],
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/core"
    },
    "Task": {
      "type": "object",
      "title": "Task represents Snap task definition.",
      "properties": {
        "creation_timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreationTimestamp"
        },
        "deadline": {
          "type": "string",
          "x-go-name": "Deadline"
        },
        "failed_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "FailedCount"
        },
        "hit_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "HitCount"
        },
        "href": {
          "type": "string",
          "x-go-name": "Href"
        },
        "id": {
          "type": "string",
          "uniqueItems": true,
          "x-go-name": "ID"
        },
        "last_failure_message": {
          "type": "string",
          "x-go-name": "LastFailureMessage"
        },
        "last_run_timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LastRunTimestamp"
        },
        "miss_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "MissCount"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "schedule": {
          "x-go-name": "Schedule",
          "$ref": "#/definitions/Schedule"
        },
        "task_state": {
          "type": "string",
          "x-go-name": "State"
        },
        "workflow": {
          "x-go-name": "Workflow",
          "$ref": "#/definitions/WorkflowMap"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/mgmt/rest/v2"
    },
    "Tasks": {
      "description": "Tasks a slice of Task",
      "type": "array",
      "items": {
        "$ref": "#/definitions/Task"
      },
      "x-go-package": "github.com/intelsdi-x/snap/mgmt/rest/v2"
    },
    "WorkflowMap": {
      "description": "WorkflowMap A map of a desired workflow that is used to create a scheduleWorkflow",
      "type": "object",
      "required": [
        "collect"
      ],
      "properties": {
        "collect": {
          "x-go-name": "Collect",
          "$ref": "#/definitions/CollectWorkflowMapNode"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/scheduler/wmap"
    }
  },
  "responses": {
    "ErrorResponse": {
      "description": "Error unsuccessful generic response to a failed API call",
      "schema": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      }
    },
...
    "TaskErrorResponse": {
      "description": "RemoveTaskError unsuccessful generic response to a failed API call"
    },
    "TaskResponse": {
      "description": "TaskResp returns the giving task.",
      "schema": {
        "$ref": "#/definitions/Task"
      }
    },
    "TasksResponse": {
      "description": "TasksResp returns a list of created tasks.",
      "schema": {
        "type": "object",
        "properties": {
          "tasks": {
            "x-go-name": "Tasks",
            "$ref": "#/definitions/Tasks"
          }
        }
      }
    }
  }
}